+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
-#### 3.0.0 - Tuesday, October 16, 2018
+#### 3.0.0 - Wednesday, October 17, 2018
 * Converted to .net core and fsprojects scaffold
+* Changed Builder to use BlockEx
+* Removed Obsolete attribute from Builder